First commit
The first part of task 2.1 "implement own SOAP Web Service(s) with necessary supportive functionality". I managed to create a bunch of classes from a WSDL file based on the functionality of the contents of the "ws" package using wsimport. I used the front end of task 1 to test the classes nad my new servlet. For some reason the service crashes while running the content of Soap.java when trying to connect to a port to create a SOAP object.